dashboard/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
pyarrow==15.0.0
boto3==1.34.0
python-dotenv==1.0.0
